telegram-bot/pyproject.toml
[project]
name = "telegram-bot"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "dotenv>=0.9.9",
    "grpcio-tools>=1.73.1",
    "openai>=1.97.0",
    "protobuf>=6.31.1",
    "pydantic>=2.11.7",
    "python-telegram-bot>=22.2",
]
